+++ Makefile
@@ -6,7 +6,7 @@
 #    By: tparratt <tparratt@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/08/26 15:24:40 by tparratt          #+#    #+#              #
-#    Updated: 2024/09/12 13:15:19 by tparratt         ###   ########.fr        #
+#    Updated: 2024/09/12 13:27:08 by tparratt         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -19,6 +19,7 @@ NAME		=	cub3d
 SRC		=	srcs/main.c \
 			srcs/validation.c \
 			srcs/utils.c \
+			srcs/utils2.c \
 			srcs/parsing.c \
 			srcs/parsing2.c \
 			srcs/parsing3.c \
